- Removed legacy .Net 4 test project - Added .net core 2.0.0 test project - Ported legacy classes from old project to new project
