+++ SimpleMapper.Test/SimpleMapper.Test.csproj
@@ -52,24 +52,7 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ConversionCopyIntegrationTests.cs" />
-    <Compile Include="CopyIntegrationTests.cs" />
-    <Compile Include="Models\ConvertCopy\Bar.cs" />
-    <Compile Include="Models\ConvertCopy\Box.cs" />
-    <Compile Include="Models\ConvertCopy\Foo.cs" />
-    <Compile Include="Models\ConvertCopy\FooMapFrom.cs" />
-    <Compile Include="Models\ConvertCopy\BarParse.cs" />
-    <Compile Include="Models\ConvertCopy\FooParse.cs" />
-    <Compile Include="Models\ConvertCopy\FooRejectNullReferences.cs" />
-    <Compile Include="Models\ConvertCopy\FooRequireAllProperties.cs" />
-    <Compile Include="Models\DefaultCopy\Bar.cs" />
-    <Compile Include="Models\DefaultCopy\Box.cs" />
-    <Compile Include="Models\DefaultCopy\Foo.cs" />
-    <Compile Include="Models\DefaultCopy\FooMapFrom.cs" />
-    <Compile Include="Models\DefaultCopy\FooRejectNullReferences.cs" />
-    <Compile Include="Models\DefaultCopy\FooRequireAllProperties.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ValidatorTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
