+++ Tests/Models/ConvertCopy/Bar.cs
@@ -1,16 +1,12 @@
-﻿using SimpleMapper;
-using SimpleMapper.Attributes;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test.Models.ConvertCopy
+namespace Tests.Models.ConvertCopy
 {
-    [ImplicitConversionAttribute]
-    [RequireAllProperties]
-    public class FooRequireAllProperties : BaseMappable<Bar, FooRequireAllProperties>
+    public class Bar
     {
         public string FirstName { get; set; }
         public int Age { get; set; }
@@ -20,12 +16,5 @@ namespace ShipBob.Mapper.Test.Models.ConvertCopy
         public DateTime Birthdate { get; set; }
         public TimeSpan TimeSpanAlive { get; set; }
         public bool IsHappy { get; set; }
-
-
-
-
-
-        //This property is not on the Bar class
-        public bool RandomProperty { get; set; }
     }
 }
