+++ Tests/Models/ConvertCopy/FooRequireAllProperties.cs
@@ -6,18 +6,26 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test.Models.ConvertCopy
+namespace Tests.Models.ConvertCopy
 {
     [ImplicitConversionAttribute]
-    public class FooParse : BaseMappable<BarParse, FooParse>
+    [RequireAllProperties]
+    public class FooRequireAllProperties : BaseMappable<Bar, FooRequireAllProperties>
     {
         public string FirstName { get; set; }
         public int Age { get; set; }
         public long SecondsAlive { get; set; }
         public decimal MoneyToTheWallet { get; set; }
         public double GPA { get; set; }
-        public string Birthdate { get; set; }
+        public DateTime Birthdate { get; set; }
         public TimeSpan TimeSpanAlive { get; set; }
         public bool IsHappy { get; set; }
+
+
+
+
+
+        //This property is not on the Bar class
+        public bool RandomProperty { get; set; }
     }
 }
