+++ Tests/Models/DefaultCopy/Bar.cs
@@ -4,15 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test.Models.ConvertCopy
+namespace Tests.Models.DefaultCopy
 {
-    public class BarParse
+    public class Bar
     {
         public string FirstName { get; set; }
-        public string Age { get; set; }
-        public string SecondsAlive { get; set; }
-        public string MoneyToTheWallet { get; set; }
-        public string GPA { get; set; }
+        public int Age { get; set; }
+        public long SecondsAlive { get; set; }
+        public decimal MoneyToTheWallet { get; set; }
+        public double GPA { get; set; }
         public DateTime Birthdate { get; set; }
         public TimeSpan TimeSpanAlive { get; set; }
         public bool IsHappy { get; set; }
