+++ Tests/Models/DefaultCopy/FooRejectNullReferences.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test.Models.DefaultCopy
+namespace Tests.Models.DefaultCopy
 {
     [RejectNullReferences]
     public class FooRejectNullReferences : BaseMappable<Bar, FooRejectNullReferences>
