+++ Tests/Models/DefaultCopy/FooRequireAllProperties.cs
@@ -6,13 +6,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test.Models.ConvertCopy
+namespace Tests.Models.DefaultCopy
 {
-    [ImplicitConversionAttribute]
-    [RejectNullReferences]
-    public class FooRejectNullReferences : BaseMappable<Bar, FooRejectNullReferences>
+    [RequireAllProperties]
+    public class FooRequireAllProperties : BaseMappable<Bar, FooRequireAllProperties>
     {
-        [RejectNullReferences]
         public string FirstName { get; set; }
         public int Age { get; set; }
         public long SecondsAlive { get; set; }
@@ -21,5 +19,12 @@ namespace ShipBob.Mapper.Test.Models.ConvertCopy
         public DateTime Birthdate { get; set; }
         public TimeSpan TimeSpanAlive { get; set; }
         public bool IsHappy { get; set; }
+
+
+
+
+
+        //This property is not on the Bar class
+        public bool RandomProperty { get; set; }
     }
 }
