+++ Tests/ValidatorTests.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ShipBob.Mapper.Test
+namespace Tests
 {
     [TestClass]
     public class ValidatorTests
